+++ test/test_helper.exs
@@ -35,7 +35,8 @@ ExUnit.start(
     # :issuing_transaction_log,
     # :issuing_withdrawal,
     # :issuing_withdrawal_log,
-    # :webhook
+    # :webhook,
+    # :brcode_preview
   ]
 )
 
@@ -48,3 +49,4 @@ Code.require_file("./test/utils/random.exs")
 Code.require_file("./test/utils/issuing_holder.exs")
 Code.require_file("./test/utils/issuing_withdrawal.exs")
 Code.require_file("./test/utils/issuing_invoice.exs")
+Code.require_file("./test/utils/static_brcode.exs")